Attempt to test log4sh, could not get working in cygwin
Attempt to organize the main script to test the logging framework, but I could not get it to work within cygwin. Complaints about a -r flag not being understood. Will need to test on a Linux machine to be sure.

Added a config file to see if the loading was the issue.