Refactor Directory Structure so all things code related are placed under the source directory.
By including all folders related to code in the source directory it allows better inclusion paths by not having to have to traverse up the directory path to the root. It also makes it local to the entry-point shell application. Tests can also be localized, but may require traversing up, which is better than the live application. Although the test folder may move back up at a later date.